More Littleroot Scripting
Scripting is nearly complete. Bag needs to be removed once starter is picked. Rival needs some text before walking to route one. Rival's mom's house needs to not have the scripting for it. NPCs inside of your house need to be removed, so does the pokeball on the second floor.